+++ docs/development/releasenotes.rst
@@ -8,6 +8,9 @@ Tags
 ~~~~
 1. *Tables*. Tags to render tables.
 2. *Data*. Tags to incorporate data into tables or plots.
+3. *Captions*. Captions now render in html using the ``caption`` or
+   ``figcaption`` tags.
+4. *Figures*. Figures now render in html using the ``figure`` tag.
 
 v0.19
 -----